refactor/Change the location of the BaseEntity
