+++ src/main/java/hello/newsfeed/NewsfeedApplication.java
@@ -2,7 +2,9 @@ package hello.newsfeed;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class NewsfeedApplication {
